Make Scheduler#sleep an async method
This makes it more consistent with the behavior one may expect from calling an "async function", that no code will be executed until the awaitable is submitted to the loop. Previously this method looks like it's "async" but it's actually not, and part of it gets executed immediately even before any awaitable is returned. So e.g. submitting it directly to Scheduler#run_coroutine() will mask a thread immediately and cause the method to seemingly behave in an unexpected way.